# go-gorilla-mux/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "電影、書籍和帖子的綜合 RESTful API",
        "title": "整合 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API 支援",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/movies": {
            "get": {
                "description": "獲取系統中所有電影的列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "獲取所有電影",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Movie"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "創建新的電影記錄",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "創建電影",
                "parameters": [
                    {
                        "description": "電影資料",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    },
                    "400": {
                        "description": "無效的請求體",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "根據 ID 獲取特定電影的詳細信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "獲取特定電影",
                "parameters": [
                    {
                        "type": "string",
                        "description": "電影 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    },
                    "404": {
                        "description": "電影未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "根據 ID 更新特定電影的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "更新電影",
                "parameters": [
                    {
                        "type": "string",
                        "description": "電影 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "電影資料",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    },
                    "400": {
                        "description": "無效的請求體",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "電影未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "根據 ID 刪除特定電影",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "刪除電影",
                "parameters": [
                    {
                        "type": "string",
                        "description": "電影 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "電影未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Director": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string",
                    "example": "Steven"
                },
                "lastname": {
                    "type": "string",
                    "example": "Spielberg"
                }
            }
        },
        "models.Movie": {
            "type": "object",
            "properties": {
                "director": {
                    "$ref": "#/definitions/models.Director"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "isbn": {
                    "type": "string",
                    "example": "1234567890"
                },
                "title": {
                    "type": "string",
                    "example": "電影標題"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}